refactor: move the journal beside its consumers, and re-home interp's JIT tests
internal/jit/journal sat under a package that never imported it. Its consumers
are internal/jit/arm64, which emits accesses to it, and interp, which owns the
buffer and reads exit state back - the compiler in between does not touch it.
Nesting it there asserted an ownership that did not exist, which is the shape
§6 names: split into a sibling package, not a nested implementation hierarchy.
It is internal/journal now, a leaf below both. Package name is unchanged, so
every call site reads the same.

interp's JIT files were inconsistently named and its test files were mis-owned.
The selector is jit_arm64.go and jit_stub.go now, not native_*. jit_test.go had
no jit.go to own it since that driver moved out, so its black-box parity suite
joins interp_test.go, which owns the Run contract those tests extend.
native_test.go matched the selector by name while testing none of it; it splits
by what each test actually reads - TestARM64_StackReserve to tier_test.go,
where nativeFrameLimit is declared, and the rest to compile_test.go.

Two of TestCompiler_Compile's subtests turned out to construct an Interpreter
and a compiler they never used: they only drove jit.StaticPlan. They are
internal/jit's own tests now, with the dead setup gone.
