+++ docs/testing.md
@@ -38,13 +38,14 @@ than widening production APIs for test access.
 
 ### White-Box Test Files
 
-Two test files still declare `package interp` because the contracts they
+Three test files still declare `package interp` because the contracts they
 assert have no external-package expression yet. Each is recorded with the
 condition that removes it (`docs/coding-patterns.md` §1.1).
 
 | File | Why it is white-box | Removal condition |
 |---|---|---|
-| `interp/native_test.go` | Drives `jit.Compiler.Compile` and calls the compiled callable directly with `journalPtr`, asserting the native trap encoding (`journal.CellTrap`, `journal.CellExitID`, `jit.ExitDescriptor`); splices frame state to force a specific exit; reads install bookkeeping (`tried`, `exits`) that no metric separates from "not attempted"; and ties private `nativeFrameLimit` to the ARM64 trampoline's stack reserve via `arm64.StackReserve`, which owns the byte arithmetic (`internal/asm/arm64/stack.go`). The complementary half of that reserve invariant — that `abi_arm64.s`'s own two literals agree with each other — needs no private interp state and lives as an ordinary external test, `arm64.TestFrameSize` (`internal/asm/arm64/stack_test.go`). | None known. The ARM64 backend itself moved to `internal/jit/arm64`, but `newCompiler`, `compileSnapshot`, `tried`, `exits`, and `nativeFrameLimit` are interp-private bookkeeping and constants with no contract to move with it. |
+| `interp/tier_test.go` | Ties private `nativeFrameLimit` (declared in `interp/tier.go`) to the ARM64 invoke trampoline's stack reserve and total frame size via `arm64.StackReserve`/`arm64.FrameSize`, which own the byte arithmetic (`internal/asm/arm64/stack.go`), and reads `abi_arm64.s`'s two literals directly to compare against them. The complementary half of that reserve invariant — that `abi_arm64.s`'s own two literals agree with each other — needs no private interp state and lives as an ordinary external test, `arm64.TestFrameSize` (`internal/asm/arm64/stack_test.go`). | None known. `nativeFrameLimit` is interp-private bookkeeping with no contract to move with the ARM64 backend, which already lives in `internal/jit/arm64`. |
+| `interp/compile_test.go` | Drives `jit.Compiler.Compile` against a snapshot only the private `Interpreter.compileSnapshot` can build (module, constants, globals, heap, declared types, and the `HostStruct`/`field`/`conversion`/`coroutine` layout offsets `jit.Layout` needs), installs the result with the private `Interpreter.install`, and calls the compiled callable directly with `journalPtr`, asserting the native trap encoding (`journal.CellTrap`, `journal.CellExitID`, `jit.ExitDescriptor`) and splicing frame state to force a specific exit; also reads install bookkeeping (`tried`, `exits`) and `tracer.headers` that no metric separates from "not attempted" (`TestARM64_Backedge`). | None known. `newCompiler`, `compileSnapshot`, `journalPtr`, `install`, `tried`, `exits`, and the loop/call dispatch wrappers are interp-private mechanics with no contract to move with the ARM64 backend, which already lives in `internal/jit/arm64`. |
 | `interp/trace_test.go` | Calls `tracer.capture` without `Run`, which is the only way to prove speculative capture snapshots a container instead of mutating the live heap. Any public path runs the real instructions too, so the isolation claim is unobservable by construction. | None known. The recorder stays in `interp`, and the claim is only provable from inside it. |
 
 ### Known Coverage Gaps
