+++ internal/asm/arm64/stack.go
@@ -15,7 +15,7 @@ const SaveAreaBytes = 80
 // StackReserve returns the native stack reserve invoke must declare: room
 // for asm.MaxSpillSlots spill slots plus callDepth native call-depth
 // records of recordBytes each. The caller supplies recordBytes (see
-// journal.Shift) rather than this package importing internal/jit/journal,
+// journal.Shift) rather than this package importing internal/journal,
 // keeping architecture encoding free of a dependency on the frame-journal
 // layout it has no other reason to know about.
 func StackReserve(recordBytes, callDepth int) int {
