+++ interp/jit_stub.go
