+++ interp/tier.go
@@ -30,7 +30,7 @@ type watchdog struct {
 
 // nativeFrameLimit caps generated call depth to the stack space reserved by
 // the ARM64 invoke trampoline (see arm64.StackReserve and
-// TestARM64_StackReserve in native_test.go). Deeper calls trap before moving
+// TestARM64_StackReserve in tier_test.go). Deeper calls trap before moving
 // SP.
 const nativeFrameLimit = 128
